+++ edge/pkg/metamanager/process.go
@@ -16,6 +16,7 @@ import (
 	"github.com/kubeedge/kubeedge/common/constants"
 	connect "github.com/kubeedge/kubeedge/edge/pkg/common/cloudconnection"
 	"github.com/kubeedge/kubeedge/edge/pkg/common/modules"
+	"github.com/kubeedge/kubeedge/edge/pkg/metamanager/client"
 	metaManagerConfig "github.com/kubeedge/kubeedge/edge/pkg/metamanager/config"
 	"github.com/kubeedge/kubeedge/edge/pkg/metamanager/dao"
 	"github.com/kubeedge/kubeedge/edge/pkg/metamanager/metaserver/kubernetes/storage/sqlite/imitator"
@@ -333,21 +334,6 @@ func processDeletePodDB(message model.Message) error {
 	return nil
 }
 
-// KeyFunc keys should be nonconfidential and safe to log
-func KeyFunc(name, namespace string, tr *authenticationv1.TokenRequest) string {
-	var exp int64
-	if tr.Spec.ExpirationSeconds != nil {
-		exp = *tr.Spec.ExpirationSeconds
-	}
-
-	var ref authenticationv1.BoundObjectReference
-	if tr.Spec.BoundObjectRef != nil {
-		ref = *tr.Spec.BoundObjectRef
-	}
-
-	return fmt.Sprintf("%q/%q/%#v/%#v/%#v", name, namespace, tr.Spec.Audiences, exp, ref)
-}
-
 // getSpecialResourceKey get service account db key
 func getSpecialResourceKey(resType, resKey string, message model.Message) (string, error) {
 	if resType != model.ResourceTypeServiceAccountToken {
@@ -361,7 +347,7 @@ func getSpecialResourceKey(resType, resKey string, message model.Message) (strin
 	if len(tokens) != 3 {
 		return "", fmt.Errorf("failed to get resource %s name and namespace", resKey)
 	}
-	return KeyFunc(tokens[2], tokens[0], tokenReq), nil
+	return client.KeyFunc(tokens[2], tokens[0], tokenReq), nil
 }
 
 func (m *metaManager) processQuery(message model.Message) {
